requirements.txt
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
alembic
amqp
annotated-types
antlr4-python3-runtime
anyio
appdirs
async-timeout
asyncssh
atpublic
attrs
Authlib
billiard
blinker
cachetools
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
cloudpickle
colorama
configobj
contourpy
cryptography
cycler
cyclopts
databricks-sdk
debugpy
decorator
dictdiffer
diskcache
distro
dnspython
docker
docstring_parser
docutils
dpath
dulwich
dvc
dvc-data
dvc-http
dvc-objects
dvc-render
dvc-studio-client
dvc-task
email-validator
entrypoints
exceptiongroup
executing
fastapi
fastmcp
filelock
Flask
flatten-dict
flufl.lock
fonttools
frozenlist
fsspec
funcy
gitdb
GitPython
google-auth
grandalf
graphene
graphql-core
graphql-relay
greenlet
gto
h11
httpcore
httpx
httpx-sse
hydra-core
idna
imbalanced-learn
imblearn
importlib_metadata
iniconfig
isodate
iterative-telemetry
itsdangerous
Jinja2
joblib
jsonschema
jsonschema-path
jsonschema-specifications
kiwisolver
kombu
keras
lazy-object-proxy
Mako
markdown-it-py
MarkupSafe
matplotlib
mcp
mdurl
mlflow
mlflow-skinny
mlflow-tracing
ml-dtypes
more-itertools
multidict
nest-asyncio
networkx
numpy==1.26.4
omegaconf
openapi-core
openapi-pydantic
openapi-schema-validator
openapi-spec-validator
opentelemetry-api
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
packaging
pandas
parse
parso
pathable
pathspec
pillow
platformdirs
pluggy
prompt_toolkit
propcache
protobuf
psutil
pure_eval
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
pydot
pygit2
Pygments
pygtrie
pyparsing
pyperclip
pytest
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
pyzmq
referencing
requests
rfc3339-validator
rich
rich-rst
rpds-py
rsa
ruamel.yaml
ruamel.yaml.clib
scikit-learn
scikeras
scipy
scmrepo
seaborn
semver
shellingham
shortuuid
shtab
six
smmap
sniffio
SQLAlchemy
sqlparse
sqltrie
sse-starlette
starlette
streamlit
tabulate
tensorflow==2.20.0
tensorboard
threadpoolctl
tomli
tomlkit
tornado
tqdm
traitlets
typer
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
vine
voluptuous
waitress
wcwidth
Werkzeug
xgboost
yarl
zc.lockfile
zipp